refactor: Create database text fixture
